+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2024/03/07
+  * debuger支持load config，用于让lldb加载init命令。
+
 2024/03/06
   * 完善svg支持awtk-widget-shape(感谢陈聪提供补丁)
   * 增加函数tk\_tcp\_is\_port\_in\_use